refactor(#57): move cli command declaration into cli.rs
