+++ src/lib.rs
@@ -1,3 +1,5 @@
+#[doc(hidden)]
+pub mod cli;
 pub mod common;
 #[doc(hidden)]
 pub mod gamestate;
